Organize projects and add some documentation to fix warnings
