+++ SimpleImpersonation.sln
@@ -3,11 +3,11 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.27703.2035
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SimpleImpersonation", "src\SimpleImpersonation.csproj", "{08800994-1B14-4806-BEF2-9F417AC42A76}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimpleImpersonation.UnitTests", "test\SimpleImpersonation.UnitTests\SimpleImpersonation.UnitTests.csproj", "{EC2F34BD-5A5F-415E-9D10-8F812EA170D1}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimpleImpersonation.Samples", "SimpleImpersonation.Samples\SimpleImpersonation.Samples.csproj", "{D09B65AC-B1B0-49F1-8295-B3775F4C43FA}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimpleImpersonation", "src\SimpleImpersonation\SimpleImpersonation.csproj", "{DCFF179B-CED3-42C9-B13A-6C2496D61476}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimpleImpersonation.Samples", "src\SimpleImpersonation.Samples\SimpleImpersonation.Samples.csproj", "{F3EBD76B-D6B8-488A-B289-EFB1D61D3E14}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,18 +15,18 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{08800994-1B14-4806-BEF2-9F417AC42A76}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{08800994-1B14-4806-BEF2-9F417AC42A76}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{08800994-1B14-4806-BEF2-9F417AC42A76}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{08800994-1B14-4806-BEF2-9F417AC42A76}.Release|Any CPU.Build.0 = Release|Any CPU
 		{EC2F34BD-5A5F-415E-9D10-8F812EA170D1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{EC2F34BD-5A5F-415E-9D10-8F812EA170D1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EC2F34BD-5A5F-415E-9D10-8F812EA170D1}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EC2F34BD-5A5F-415E-9D10-8F812EA170D1}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D09B65AC-B1B0-49F1-8295-B3775F4C43FA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D09B65AC-B1B0-49F1-8295-B3775F4C43FA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D09B65AC-B1B0-49F1-8295-B3775F4C43FA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D09B65AC-B1B0-49F1-8295-B3775F4C43FA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DCFF179B-CED3-42C9-B13A-6C2496D61476}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DCFF179B-CED3-42C9-B13A-6C2496D61476}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DCFF179B-CED3-42C9-B13A-6C2496D61476}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DCFF179B-CED3-42C9-B13A-6C2496D61476}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F3EBD76B-D6B8-488A-B289-EFB1D61D3E14}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F3EBD76B-D6B8-488A-B289-EFB1D61D3E14}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F3EBD76B-D6B8-488A-B289-EFB1D61D3E14}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F3EBD76B-D6B8-488A-B289-EFB1D61D3E14}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
