+++ src/SimpleImpersonation.Samples/Program.cs
