+++ src/SimpleImpersonation.Samples/SimpleImpersonation.Samples.csproj
@@ -7,7 +7,7 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\src\SimpleImpersonation.csproj" />
+      <ProjectReference Include="..\SimpleImpersonation\SimpleImpersonation.csproj" />
     </ItemGroup>
 
 </Project>
