+++ src/SimpleImpersonation/Compatibility.cs
