+++ src/SimpleImpersonation/ImpersonationException.cs
