+++ src/SimpleImpersonation/LogonType.cs
