+++ src/SimpleImpersonation/NativeMethods.cs
