+++ src/SimpleImpersonation/UserCredentials.cs
